# swagger.yaml
swagger: "2.0"
info:
  description: "This is the swagger doc for vanDevTimes"
  version: "1.0.0"
  title: "Swagger vanDevTimes"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "vantimes.info@gmail.com"
basePath: "/api"
paths:
...
  /scheduler:
    get:
      tags:
        - "Scheduler"
      summary: "Get scheduler settings of all tasks"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Scheduler"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /user:
    get:
      tags:
        - "User"
      summary: "get user account info"
      consumes:
        - "application/json"
      parameters:
        -  in: "body"
           name: "body"
           description: "a user email"
           schema:
             example: {
               "email": "regularUser@email.com"
             }
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /users:
    get:
      tags:
        - "User"
      summary: "get all users"

      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ArrayOfUsers"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
...
  /company:
    put:
      tags:
        - "Company"
      summary: "set a user's company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a user email and new company"
          schema:
            example: {
              "newComp": "ABC",
              "email": "regularUser@email.com"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "newComp": "ABC",
              "email": "regularUser@email.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
    post:
      tags:
        - "Company"
      summary: "add a company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a new company"
          schema:
            example: {
              "companyName": "aCompany"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "companyName": "aCompany"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"

    delete:
      tags:
        - "Company"
      summary: "delete a company"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "an existing company"
          schema:
            example: {
              "companyName": "Company2Delete"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "companyName": "Company2Delete"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /user/subscription:
    put:
      tags:
        - "User"
      summary: "change a user's subscription status"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "a user email and new status"
          schema:
            example: {
              "isSubscribed": true,
              "email": "regularUser@email.com"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "isSubscribed": true,
              "email": "regularUser@email.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /companies:
    get:
      tags:
        - "Company"
      summary: "get all companies"


      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "allCompanies": [
                "Galvanize",
                "Hootsuite",
                "Later"
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /updateSettings:
    put:
      tags:
        - "Settings"
      summary: "Update system settings"
      parameters:
      - in: "body"
        name: "body"
        description: "change system settings for both scheduler and permission"
        schema:
          $ref: "#/definitions/Setting"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          content:
            application/json:
              example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /permission:
    get:
      tags:
        - "Permission"
      summary: "Get permission settings of all user types"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Permission"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
definitions:
...
  User:
    type: "object"
    required:
      - "userName"
      - "email"
      - "type"
      - "companyName"
      - "subscribed"
    properties:
      userName:
        example: "John Doe"
      email:
        example: "regularUser@gmail.com"
      type:
        example: "regular"
      companyName:
        example: "Galvanize"
      subscribed:
        example: 1
  ArrayOfUsers:
    type: "object"
    required:
      - "users"
    properties:
      users:
        example: [
        {
          "USER_NAME": "Jackson Doe",
          "EMAIL": "adminUser@email.com",
          "TYPE": "admin",
          "companyName": null,
          "SUBSCRIBED": 1
        },
        {
          "USER_NAME": "Hee Su Kim",
          "EMAIL": "kessris@gmail.com",
          "TYPE": "sysManager",
          "companyName": null,
          "SUBSCRIBED": 1
        }]
  Scheduler:
    type: "object"
    properties:
      taskId:
        example: "RSS retrieval"
      newfreq:
        example: 7
      maxNum:
        example: 10
  Permission:
    type: "object"
    required:
      - "permissionType"
      - "approvalThreshold"
    properties:
      permissionType:
        example: "regular"
      approvalThreshold:
        example: 4
  Setting:
    type: "object"
    required:
      - "permission"
      - "scheduler"
    properties:
      permission:
        example: [{"newThreshold": 27, "permissionType": "regular"},{"newThreshold": 2, "permissionType": "admin"}]
      scheduler:
        example: [{"taskId": "approval-reminder", "newFreq": 3, "maxNum": 17}, {"taskId": "subscription-email", "newFreq": 1, "maxNum": 17}]
...